content/restart-content-using-airflow/app/requirements.txt
matplotlib
pandas
pins
posit-sdk
seaborn
shiny
